feat: service table & group nav on quota overview
