+++ plugins/services/src/js/components/ServicesQuotaOverviewTable.tsx
@@ -1,62 +1,34 @@
 import React from "react";
-import { Trans, Plural } from "@lingui/macro";
-import {
-  Column,
-  Table,
-  SortableHeaderCell,
-  SpacingBox,
-  InfoBoxInline,
-  Icon
-} from "@dcos/ui-kit";
-import { SystemIcons } from "@dcos/ui-kit/dist/packages/icons/dist/system-icons-enum";
-import { iconSizeXs } from "@dcos/ui-kit/dist/packages/design-tokens/build/js/designTokens";
+import { Trans } from "@lingui/macro";
+import { Column, Table, SortableHeaderCell } from "@dcos/ui-kit";
 import sort from "array-sort";
 
-import Loader from "#SRC/js/components/Loader";
+import Service from "../structs/Service";
+import Pod from "../structs/Pod";
+import ServiceTree from "../structs/ServiceTree";
 
-import { ServiceGroup } from "../types/ServiceGroup";
-import { nameRenderer } from "../columns/QuotaOverviewNameColumn";
-import { limitRenderer } from "../columns/QuotaOverviewLimitColumn";
-import { cpuRenderer } from "../columns/QuotaOverviewCPUConsumedColumn";
-import { memRenderer } from "../columns/QuotaOverviewMemoryConsumedColumn";
-import { diskRenderer } from "../columns/QuotaOverviewDiskConsumedColumn";
-import { gpuRenderer } from "../columns/QuotaOverviewGPUConsumedColumn";
+import { nameRenderer } from "../columns/ServicesTableNameColumn";
 
 import { SortDirection } from "../types/SortDirection";
 
 export interface ServicesQuotaOverviewTableProps {
-  groups: ServiceGroup[];
+  serviceTree: ServiceTree;
 }
 
-interface ServicesQuotaOverViewTableState {
-  groups: ServiceGroup[];
+type QuotaThing = Service | Pod | ServiceTree;
+
+interface ServicesQuotaOverviewTableState {
+  items: QuotaThing[];
   sortDirection: SortDirection;
   sortColumn: string;
 }
 
-const compatatorFor = (prop: string) => (a: ServiceGroup, b: ServiceGroup) =>
-  ServiceGroup.getQuotaPercentage(a, prop) -
-  ServiceGroup.getQuotaPercentage(b, prop);
-
 function sortForColumn(
   columnName: string
-): (a: ServiceGroup, b: ServiceGroup) => number {
+): (a: QuotaThing, b: QuotaThing) => number {
   switch (columnName) {
     case "name":
-      return (a, b) => a.name.localeCompare(b.name);
-    case "limit":
-      return (a, b) =>
-        (a.quota ? a.quota.limitStatus : "N/A").localeCompare(
-          b.quota ? b.quota.limitStatus : "N/A"
-        );
-    case "cpus":
-      return compatatorFor("cpus");
-    case "mem":
-      return compatatorFor("memory");
-    case "disk":
-      return compatatorFor("disk");
-    case "gpus":
-      return compatatorFor("gpus");
+      return (a, b) => a.getName().localeCompare(b.getName());
     default:
       return () => 0;
   }
@@ -64,20 +36,22 @@ function sortForColumn(
 
 class ServicesQuotaOverviewTable extends React.Component<
   ServicesQuotaOverviewTableProps,
-  ServicesQuotaOverViewTableState
+  ServicesQuotaOverviewTableState
 > {
   constructor(props: Readonly<ServicesQuotaOverviewTableProps>) {
     super(props);
 
     this.state = {
-      groups: [],
+      items: props.serviceTree.getItems(),
       sortColumn: "name",
       sortDirection: "ASC"
     };
   }
 
   UNSAFE_componentWillReceiveProps(nextProps: ServicesQuotaOverviewTableProps) {
-    this.setState({ groups: this.sortData(nextProps.groups || []) });
+    this.setState({
+      items: this.sortData(nextProps.serviceTree.getItems() || [])
+    });
   }
 
   handleSortClick = (columnName: string) => () => {
@@ -87,68 +61,31 @@ class ServicesQuotaOverviewTable extends React.Component<
         : "ASC";
 
     this.setState({
-      groups: this.sortData(this.state.groups, columnName, toggledDirection),
+      items: this.sortData(this.state.items, columnName, toggledDirection),
       sortColumn: columnName,
       sortDirection: toggledDirection
     });
   };
 
   sortData = (
-    groups: ServiceGroup[],
+    items: QuotaThing[],
     sortColumn: string = this.state.sortColumn,
     sortDirection: SortDirection = this.state.sortDirection
-  ): ServiceGroup[] =>
-    sort(groups.slice(), sortForColumn(sortColumn), {
+  ): QuotaThing[] =>
+    sort(items.slice(), sortForColumn(sortColumn), {
       reverse: sortDirection !== "ASC"
     });
 
-  getNoLimitInfobox() {
-    const { groups } = this.state;
-    const noLimitGroups = groups.filter(
-      group => group.quota && group.quota.limitStatus !== "Enforced"
-    );
+  render() {
+    const { items, sortColumn, sortDirection } = this.state;
 
-    if (!noLimitGroups.length) {
+    if (items.length === 0) {
       return null;
     }
 
     return (
-      <SpacingBox side="bottom" spacingSize="l">
-        <InfoBoxInline
-          appearance="default"
-          message={
-            <React.Fragment>
-              <Icon
-                shape={SystemIcons.CircleInformation}
-                size={iconSizeXs}
-                color="currentColor"
-              />
-              <Plural
-                render={<span id="quota-no-limit-infobox" />}
-                value={noLimitGroups.length}
-                one={`# group has services not limited by quota. Update service roles
-                to have quota enforced.`}
-                other={`# groups have services not limited by
-                quota. Update service roles to have quota enforced.`}
-              />
-            </React.Fragment>
-          }
-        />
-      </SpacingBox>
-    );
-  }
-
-  render() {
-    const { groups, sortColumn, sortDirection } = this.state;
-
-    if (!groups.length) {
-      return <Loader />;
-    }
-
-    return (
-      <div className="table-wrapper quota-table">
-        {this.getNoLimitInfobox()}
-        <Table data={groups}>
+      <div className="table-wrapper service-quota-table">
+        <Table data={items}>
           <Column
             key="name"
             header={
@@ -158,62 +95,7 @@ class ServicesQuotaOverviewTable extends React.Component<
                 sortDirection={sortColumn === "name" ? sortDirection : null}
               />
             }
-            cellRenderer={nameRenderer}
-          />
-          <Column
-            key="limit"
-            header={
-              <SortableHeaderCell
-                columnContent={<Trans render="span">Quota Limit</Trans>}
-                sortHandler={this.handleSortClick("limit")}
-                sortDirection={sortColumn === "limit" ? sortDirection : null}
-              />
-            }
-            cellRenderer={limitRenderer}
-          />
-          <Column
-            key="cpus"
-            header={
-              <SortableHeaderCell
-                columnContent={<Trans render="span">CPU Consumed</Trans>}
-                sortHandler={this.handleSortClick("cpus")}
-                sortDirection={sortColumn === "cpus" ? sortDirection : null}
-              />
-            }
-            cellRenderer={cpuRenderer}
-          />
-          <Column
-            key="mem"
-            header={
-              <SortableHeaderCell
-                columnContent={<Trans render="span">Memory Consumed</Trans>}
-                sortHandler={this.handleSortClick("mem")}
-                sortDirection={sortColumn === "mem" ? sortDirection : null}
-              />
-            }
-            cellRenderer={memRenderer}
-          />
-          <Column
-            key="disk"
-            header={
-              <SortableHeaderCell
-                columnContent={<Trans render="span">Disk Consumed</Trans>}
-                sortHandler={this.handleSortClick("disk")}
-                sortDirection={sortColumn === "disk" ? sortDirection : null}
-              />
-            }
-            cellRenderer={diskRenderer}
-          />
-          <Column
-            key="gpus"
-            header={
-              <SortableHeaderCell
-                columnContent={<Trans render="span">GPU Consumed</Trans>}
-                sortHandler={this.handleSortClick("gpus")}
-                sortDirection={sortColumn === "gpus" ? sortDirection : null}
-              />
-            }
-            cellRenderer={gpuRenderer}
+            cellRenderer={nameRenderer.bind(null, false, true)}
           />
         </Table>
       </div>
