+++ src/styles/components/quota/styles.less
@@ -30,3 +30,7 @@
     }
   }
 }
+
+.service-quota-table {
+  padding-top: @base-spacing-unit;
+}
